liked/watchlist database, liked page, like button in table

+++ src/firebase-config.js
@@ -1,7 +1,7 @@
 import { createContext, useState, useEffect, useContext } from "react";
 
 import { initializeApp } from "firebase/app";
-import { getFirestore } from "firebase/firestore";
+import { addDoc, getFirestore, onSnapshot, setDoc, doc, getDoc, collection } from "firebase/firestore";
 import { getStorage, ref, uploadBytes, getDownloadURL } from "firebase/storage";
 import {
   getAuth,
@@ -33,40 +33,54 @@ export const db = getFirestore(app);
 export const storage = getStorage(app);
 export const provider = new GoogleAuthProvider();
 
-// import this context in your component and use useContex() hook to get the value
+// import this context in your component and use useContex(FirebaseContext) hook to get the value
 export const FirebaseContext = createContext();
 
 export const FirebaseContextProvider = ({ children }) => {
+
   // this is the state that will be shared with all components
-  // so you dont need to create a new state for each component
+  // so you dont need to create a new state for each component 
   // and make additional requests to the database
 
   const [user, setUser] = useState(null);
-  const [liked_movies, setLikedMovies] = useState([]);
+  const [liked, setLikedFilms] = useState([]);
+  const [watchlist, setWatchlistFilms] = useState([]);
+  const [ratings, setRatings] = useState({});
 
   useEffect(() => {
-    onAuthStateChanged(auth, setUser);
+    onAuthStateChanged(auth, (currentUser) => {
+      setUser(currentUser)
+      onSnapshot(doc(db, "Liked", currentUser?.uid), (snapshot) => {
+        setLikedFilms(snapshot.data()?.likedFilms || []);
+      });
+      onSnapshot(doc(db, "Watchlist", currentUser?.uid), (snapshot) => {
+        setWatchlistFilms(snapshot.data()?.watchlistFilms || []);
+      });
+      onSnapshot(doc(db, "Rating", currentUser?.uid), (snapshot) => {
+        setRatings(snapshot.data()?.ratings || {});
+      });
+    })
   }, []);
 
   return (
-    <FirebaseContext.Provider value={{ user, liked_movies }}>
+    <FirebaseContext.Provider value={{ user, liked, watchlist, ratings }}> {/* <-- this is the value that will be shared with all components*/}
       {children}
     </FirebaseContext.Provider>
   );
-};
+}
 
 export const register = async (email, password, name) => {
   if (email && password && name) {
     try {
       await createUserWithEmailAndPassword(auth, email, password);
       await updateProfile(auth.currentUser, { displayName: name });
-      return true;
+      await Alert({ title: "Registration successful!", icon: "success", text: "" });
+      window.open("/", "_self");
     } catch (error) {
       Alert({ title: error.code });
     }
   } else {
-    Alert("Empty fields");
-    return false;
+    Alert({ title: "Empty fields" });
   }
 };
 
@@ -74,20 +88,38 @@ export const login = async (email, password) => {
   if (email && password) {
     try {
       await signInWithEmailAndPassword(auth, email, password);
+      await Alert({ title: "Login successful!", icon: "success", text: "" });
       window.open("/", "_self");
-      return true;
     } catch (error) {
-      return error.code;
+      Alert({ title: error.code });
     }
   } else {
-    Alert("Empty fields");
+    Alert({ title: "Empty fields" });
   }
 };
 
+export const logout = async () => {
+  try {
+    await signOut(auth);
+    window.open("/", "_self");
+  } catch (error) {
+    Alert({ title: error.code });
+  }
+}
+
 export const deleteUserAccount = async () => {
   try {
-    await deleteUser(auth.currentUser);
-    window.open("/", "_self");
+    Alert({ title: "You are going to delete your account.\n\nAre you sure?", text: "This action cannot be undone", icon: "question", timer: false, showConfirmButton: true, showCancelButton: true })
+      .then(async (result) => {
+        if (result.isConfirmed) {
+          await deleteUser(auth.currentUser);
+          await Alert({ title: "Your account succesfully deleted", icon: "success", text: "" });
+          window.open("/", "_self");
+        }
+      }).catch((error) => {
+        Alert({ title: error.code });
+      });
+
   } catch (error) {
     if (error.code === "auth/requires-recent-login") {
       Alert("You need to reauthenticate before deleting your account");
@@ -101,11 +133,23 @@ export const deleteUserAccount = async () => {
 export const signInWithGoogle = async () => {
   try {
     await signInWithPopup(auth, provider);
+    await Alert({ title: "Login successful!", icon: "success", text: "" });
     window.open("/", "_self");
   } catch (error) {
     Alert({ title: error.code });
   }
 };
+
+export const updateName = async (name) => {
+  if (name) {
+    try {
+      await updateProfile(auth.currentUser, { displayName: name });
+      window.location.reload();
+    } catch (error) {
+      Alert({ title: error.code });
+    }
+  }
+}
 
 export const uploadImage = async (file) => {
   if (auth.currentUser.uid) {
@@ -124,10 +168,10 @@ export const uploadImage = async (file) => {
         Alert({ title: error.code });
       }
     } else {
-      Alert("Please select a file");
+      Alert({ title: "Please select a file" });
     }
   } else {
-    Alert("Please login to upload a profile picture");
+    Alert({ title: "Please login to upload a profile picture" });
   }
 };
 
@@ -144,24 +188,87 @@ export const removeImage = async () => {
   }
 };
 
-export const updateName = async (name) => {
-  if (name) {
+export const like = async (filmID) => {
+  if (auth.currentUser?.uid && filmID) {
     try {
-      await updateProfile(auth.currentUser, { displayName: name });
-      window.location.reload();
-    } catch (error) {
+      const documentRef = doc(db, "Liked", auth.currentUser.uid);
+      const document = await getDoc(documentRef)
+      const likedFilms = document.data()?.likedFilms
+
+      if (likedFilms) {
+        if (likedFilms.includes(filmID)) {
+          likedFilms.splice(likedFilms?.indexOf(filmID), 1)
+          await setDoc(documentRef, { likedFilms: [...likedFilms] }, { merge: true });
+        }
+        else {
+          await setDoc(documentRef, { likedFilms: [...likedFilms, filmID] }, { merge: true });
+        }
+      }
+      else {
+        await setDoc(documentRef, { likedFilms: [filmID] }, { merge: true });
+      }
+    }
+    catch (error) {
       Alert({ title: error.code });
     }
-  } else {
-    Alert({ title: "Please enter a name" });
   }
-};
+  else {
+    Alert({ title: "Please login to like films" });
+  }
+}
 
-export const logout = async () => {
-  try {
-    await signOut(auth);
-    window.open("/", "_self");
-  } catch (error) {
-    Alert({ title: error.code });
+export const saveToWatchlist = async (filmID) => {
+  if (auth.currentUser?.uid && filmID) {
+    try {
+      const documentRef = doc(db, "Watchlist", auth.currentUser.uid);
+      const document = await getDoc(documentRef)
+      const watchlistFilms = document.data()?.watchlistFilms
+
+      if (watchlistFilms) {
+        if (watchlistFilms.includes(filmID)) {
+          watchlistFilms.splice(watchlistFilms?.indexOf(filmID), 1)
+          await setDoc(documentRef, { watchlistFilms: [...watchlistFilms] }, { merge: true });
+        }
+        else {
+          await setDoc(documentRef, { watchlistFilms: [...watchlistFilms, filmID] }, { merge: true });
+        }
+      }
+      else {
+        await setDoc(documentRef, { watchlistFilms: [filmID] }, { merge: true });
+      }
+    }
+    catch (error) {
+      Alert({ title: error.code })
+    }
   }
-};
+  else {
+    Alert({ title: "Please login to add a film to your watchlist" })
+  }
+}
+
+export const addRating = async (filmID, rating) => {
+  if (filmID && rating && auth.currentUser?.uid) {
+    try {
+      const documentRef = doc(db, "Rating", auth.currentUser.uid);
+      const document = await getDoc(documentRef)
+      const ratings = document.data()?.ratings
+
+      if (ratings) {
+        if (ratings[filmID]) {
+          ratings[filmID] = rating
+          await setDoc(documentRef, { ratings: { ...ratings } }, { merge: true });
+        }
+        else {
+          await setDoc(documentRef, { ratings: { ...ratings, [filmID]: rating } }, { merge: true });
+        }
+      }
+      else {
+        await setDoc(documentRef, { ratings: { [filmID]: rating } }, { merge: true });
+      }
+    }
+    catch (error) {
+      Alert({ title: error.code })
+    }
+  }
+
+}
